# dnn(ocl4dnn): add fusion support for Power activation and eltwise add