… modules/dnn/src/ocl4dnn/include/default_kernel_config.hpp