This PR adds fusion support for Power activation and element-wise add for ocl4dnn backend
Signed-off-by: Wu Zhiwen <zhiwen.wu@intel.com>

<!-- Please use this line to close one or multiple issues when this pullrequest gets merged
You can add another line right under the first one:
resolves #1234
resolves #1235
-->

### This pullrequest changes

<!-- Please describe what your pullrequest is changing -->