@alalek 2017-11-20
Thank you! Looks good to me :+1:

i5-6600 (24CU, Linux, SRB5):

|Name of Test|base|patch|speedup|
|---|:-:|:-:|:-:|
|AlexNet::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_CPU)|17.957 ms|18.196 ms|0.99|
|AlexNet::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_OPENCL)|17.024 ms|17.077 ms|1.00|
|ENet::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_CPU)|110.044 ms|110.154 ms|1.00|
|ENet::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_OPENCL)|47.933 ms|45.301 ms|1.06|
|GoogLeNet::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_CPU)|30.205 ms|30.086 ms|1.00|
|GoogLeNet::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_OPENCL)|21.962 ms|22.071 ms|1.00|
|Inception_5h::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_CPU)|31.821 ms|32.073 ms|0.99|
|Inception_5h::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_OPENCL)|25.908 ms|25.970 ms|1.00|
|ResNet50::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_CPU)|64.454 ms|64.413 ms|1.00|
|ResNet50::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_OPENCL)|49.338 ms|39.981 ms|1.23|
|SqueezeNet_v1_1::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_CPU)|8.631 ms|8.634 ms|1.00|
|SqueezeNet_v1_1::DNNTestNetwork::(DNN_BACKEND_DEFAULT, DNN_TARGET_OPENCL)|7.116 ms|7.090 ms|1.00|